validation and clear button fixes
I put some more validations to the form and added better alert messages.
(negative or zero values, user must be 18+ years old, only numbers are allowed in the fields.)

Also i fixed the problem when if you wanted to submit a new form you had to reload the page. Now you can just click the clear button and you are good to go.